moved register page to jsx templater
