+++ src/components/Avatar/index.ts
@@ -1,6 +1,6 @@
-import { Component } from "../../templator";
-import template from "./template.hbs";
-import noAvatarImage from "./../../shared/img/user.svg";
+import { Component } from "@/lib/templator";
+import template from "./template.tpl";
+import noAvatarImage from "@/shared/img/user.svg";
 
 type PropsType = {
     name: string;
