+++ src/pages/Register/template.tpl
@@ -1,4 +1,11 @@
-export default function template() {
+import { Form } from "@/components/Form";
+
+export default function template({
+    elements,
+    submit,
+    onFieldBlur,
+    onSubmit
+}) {
     return (
         <>
             <div className="page-register">
@@ -6,7 +13,8 @@ export default function template() {
                     <div className="title">Регистрация</div>
                 </div>
                 <div className="content">
-                    <span data-form></span>
+                    <Form elements={elements} submit={submit} onSubmit={onSubmit}
+                          onFieldBlur={onFieldBlur} />
                     <div className="register-link">
                         <a href="#">Войти</a>
                     </div>
